+++ CHANGELOG.md
@@ -30,6 +30,9 @@ _This sections lists changes committed since most recent release_
 - Unit tests for ModalDescription, ModalItem, ModalPostTags, ModalText, ModalTranscript, Popover, and Popup
 - A recent graphics section to the featured section on the homepage
 - Tests for alt text display to GraphicProject
+- Bottom buttons (i.e., Preview, Publish, Publish Changes, Unpublish) to GraphicEdit
+- The editable field to the supportFileDetails fragment
+- A refetch of the GRAPHIC_PROJECT_QUERY in GraphicEdit after adding new support files to an existing project to update the support files UI 
 
 **Changed:**
 
@@ -68,7 +71,11 @@ _This sections lists changes committed since most recent release_
 - Adjust modal stylesheets for consistency
 - Update unit tests for ActionButtons, GraphicEdit, GraphicProject, and ModalContentMeta
 - Move the fetching of featured item results and post types from Redux to Context
-- Update unit tests and mock data to address console messages
+- Update unit tests and mock data to address console messages 
+- Refactor collating of graphic project files in GraphicEdit to address clean shell files being displayed in the wrong column
+- Refactor tests for GraphicEdit and add additional mock data
+- The footer questions and feedback link to "gpadigitalhelp@state.gov" from "design@america.gov" 
+ 
 
 **Fixed:**